+++ Demo/scripts/app.js
@@ -1,8 +1,4 @@
-const {
-    Kernel,
-    Setting,
-    Sheet
-} = require("./easy-jsbox")
+const { Kernel, Setting, Sheet } = require("./libs/easy-jsbox")
 
 class AppKernel extends Kernel {
     constructor() {
@@ -61,17 +57,19 @@ module.exports = {
         } else {
             $intents.finish("不支持在此环境中运行")
             $ui.render({
-                views: [{
-                    type: "label",
-                    props: {
-                        text: "不支持在此环境中运行",
-                        align: $align.center
-                    },
-                    layout: (make, view) => {
-                        make.center.equalTo(view.super)
-                        make.size.equalTo(view.super)
+                views: [
+                    {
+                        type: "label",
+                        props: {
+                            text: "不支持在此环境中运行",
+                            align: $align.center
+                        },
+                        layout: (make, view) => {
+                            make.center.equalTo(view.super)
+                            make.size.equalTo(view.super)
+                        }
                     }
-                }]
+                ]
             })
         }
     }
